requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
alembic==1.10.4
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.5
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Session==0.4.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.4